tests: move some tests
